improve compile API

+++ packages/core/penc/src/compile.ts
@@ -14,12 +14,31 @@ import {generateTargetCode} from './transforms';
 import {AbsPath} from './utils';
 
 
-export function compile(options: CompilerOptions) {
+// TODO: rename...
+export interface CompilerResult {
+
+    // TODO: doc...
+    parse(text: string): unknown;
+
+    // TODO: doc...
+    print(ast: unknown): string;
+
+    // TODO: doc better...
+    /** Returns the target code as source. */
+    toString(): void;
+
+    // TODO: doc...
+    // Path where the output file will be generated. If the file already exists, it will be overwritten.
+    // If relative, will be resolved against the CWD.
+    // If omitted, will be same as main path but with extension changed to '.js'
+    save(filename?: string): void;
+}
+
+
+export function compile(options: CompilerOptions): CompilerResult {
 
     // Parse and validate compiler options
     const main = AbsPath(path.resolve(options.main));
-    const outFile = options.outFile || main.substr(0, main.length - path.extname(main).length) + '.js';
-    if (main === outFile) throw new Error(`output would overwrite input`);
 
     // Proceed through all stages in the compiler pipeline.
     const ast1 = parseSourceFiles({main});
@@ -27,9 +46,19 @@ export function compile(options: CompilerOptions) {
     const ast3 = normaliseExpressions(ast2);
     const consts = resolveConstantValues(ast3);
     const targetCode = generateTargetCode({ast: ast3, consts});
-
-    // write the target code to the output file path. Creating containing dirs if necessary.
-    const outFilePath = path.resolve(outFile);
-    fs.ensureDirSync(path.dirname(outFilePath));
-    fs.writeFileSync(outFilePath, targetCode);
+    const evalTarget = (): typeof target => target = eval(`(function (module) {\n${targetCode}\n})()`);
+    let target: {parse(text: string): unknown, print(ast: unknown): string};
+    return {
+        parse: text => (target || evalTarget()).parse(text),
+        print: ast => (target || evalTarget()).print(ast),
+        toString: () => targetCode,
+        save(filename) {
+            // write the target code to the output file path. Creating containing dirs if necessary.
+            const outFile = filename || main.substr(0, main.length - path.extname(main).length) + '.js';
+            if (main === outFile) throw new Error(`output would overwrite input`);
+            const outFilePath = path.resolve(outFile);
+            fs.ensureDirSync(path.dirname(outFilePath));
+            fs.writeFileSync(outFilePath, targetCode);
+        },
+    };
 }
